+++ packages/ccs-frontend/src/ccs/all.puppeteer.test.js
@@ -51,6 +51,7 @@ describe('GOV.UK Frontend', () => {
       // Ensure GOV.UK Frontend exports the expected components
       expect(components).toEqual([
         'Header',
+        'PasswordStrength',
       ])
     })
   })