+++ packages/ccs-frontend/src/ccs/all.ts
@@ -1,3 +1,4 @@
 export { version } from './common/ccs-frontend-version'
 export { Header } from './components/header/header'
+export { PasswordStrength } from './components/password-strength/password-strength'
 export { initAll, createAll } from './init'